requirements.txt
ttkbootstrap
pyautogui
opencv-python
pillow
pyperclip
pytesseract
psutil